Dataset
This dataset is exported from the Netherlands National Institute for Public Health and the Environment (RIVM). The dataset includes variables that are related to the health of the districts and neighborhoods in the Netherlands for people who are 18 years or above. The dataset provides information on exercise standards, perceived health, smoking and alcohol consumption, risk of depression, loneliness and psychological complaints that are indicative for individual health and overweight.