Dataset
This dataset is exported from the Centraal Bureau voor Statistiek (CBS) that is constructed in 2022. The dataset includes variables that are related to the demographics of the districts and neighborhoods in the Netherlands and provides information on the population, living standards, energy, education attainment, labor, income, social security, businesses, motor vehicles, services and urbanity which are directly effective on the standard of living.